fix(backend): improve IPPrefixReconcileQuery performance
By directly filtering the node labels. Using WHERE in labels() did not
trigger the NodeByLabelScan and thus processed all nodes.
Also add a WHERE IN clause for the binary address lookup so that the
binary address index is effectively used.
Looking up for the binary address + prefix length tuple prevented the
planner from using the index.

Signed-off-by: Fatih Acar <fatih@opsmill.com>