fix(backend): improve uniqueness constraint query
This should avoid processing useless rows when we have relationship only
constraints.
This especially improves the IPAM use-case, where the uniqueness
constraint is the (address/prefix, namespace) tuple.
Without this change, the query would check for nodes having that same
address and fetch ALL nodes present in the same namespace.
With this change, the query is now FIRST checking nodes having that same
address and THEN checking if these previous nodes are in the same namespace.

Signed-off-by: Fatih Acar <fatih@opsmill.com>